Move navigation manager
