+++ app/src/main/java/com/igorwojda/showcase/app/presentation/NavHostActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.igorwojda.showcase.R
-import com.igorwojda.showcase.library.base.navigation.NavigationManager
 import com.igorwojda.showcase.library.base.presentation.activity.BaseActivity
+import com.igorwojda.showcase.library.base.presentation.navigation.NavigationManager
 import kotlinx.android.synthetic.main.activity_nav_host.*
 import org.kodein.di.generic.instance
