+++ feature_album/src/main/java/com/igorwojda/showcase/feature/album/presentation/albumlist/AlbumListViewModel.kt
@@ -3,7 +3,7 @@ package com.igorwojda.showcase.feature.album.presentation.albumlist
 import androidx.lifecycle.viewModelScope
 import com.igorwojda.showcase.feature.album.domain.model.AlbumDomainModel
 import com.igorwojda.showcase.feature.album.domain.usecase.GetAlbumListUseCase
-import com.igorwojda.showcase.library.base.navigation.NavigationManager
+import com.igorwojda.showcase.library.base.presentation.navigation.NavigationManager
 import com.igorwojda.showcase.library.base.presentation.viewmodel.BaseAction
 import com.igorwojda.showcase.library.base.presentation.viewmodel.BaseViewModel
 import com.igorwojda.showcase.library.base.presentation.viewmodel.BaseViewState
