+++ library_base/src/main/java/com/igorwojda/showcase/library/base/BaseLibraryKodeinModule.kt
@@ -1,10 +1,13 @@
 package com.igorwojda.showcase.library.base
 
-import com.igorwojda.showcase.library.base.navigation.navigationModule
+import com.igorwojda.showcase.library.base.presentation.navigation.NavigationManager
 import org.kodein.di.Kodein
+import org.kodein.di.generic.bind
+import org.kodein.di.generic.singleton
 
 internal const val MODULE_NAME = "Base"
 
 val baseModule = Kodein.Module("${MODULE_NAME}Module") {
-    import(navigationModule)
+
+    bind() from singleton { NavigationManager() }
 }
